Merged PR 7160924: Allow non-string parameters
Updates the module to allow users to specify non-string parameters for New-GuestConfigurationPolicy. Parameters of type string, Boolean, integer, or double are supported. The module internally converts these values to strings so the end result, the policy definition, is unchanged.

This change accompanies another change in the agent where the agent automatically converts string parameters to the type expected by the resource. The end goal of these changes is to support resources requiring non-string parameters while maintaining a good user experience. With this, the customer can specify an integer parameter if the resource expects an integer and the module + agent will handle the rest.

Related work items: #16224786